refactor: move scraping files to services/ and fallback to repositories/
