+++ app/routes/importation.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Query, Depends, HTTPException
-from app.scraping.importation import fetch_importation_data, IMPORT_TYPE_TO_SUBOPT
+from app.services.importation import fetch_importation_data, IMPORT_TYPE_TO_SUBOPT
 from app.core.auth import get_current_user
 from app.logging.logger import setup_logger
 from app.models.importation_types import ImportTypeEnum
