+++ app/services/processing.py
@@ -4,8 +4,8 @@ from bs4 import BeautifulSoup
 
 from app.logging.logger import setup_logger
 from app.core.config import settings
-from app.scraping.helpers import parse_category_table
-from app.utils.fallback import load_processing_csv
+from app.services.helpers import parse_category_table
+from app.repositories.fallback import load_processing_csv
 from app.models.processing_types import GrapeTypeEnum
 
 logger = setup_logger(__name__)
